+++ aios_core/modules/viber/__init__.py
@@ -1,0 +1,11 @@
+"""Viber platform module (scaffolded onboarding package)."""
+
+from .bootstrap import ViberBootstrap
+from .messenger import ViberMessenger
+from .storage import ViberStorage
+
+__all__ = [
+    "ViberBootstrap",
+    "ViberMessenger",
+    "ViberStorage",
+]